requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
Faker==28.4.1
pytest==8.3.2
selenium==4.24.0
